move 4 utility classes to a more appropriate package
