+++ plugins/org.eclipse.n4js.ui/src/org/eclipse/n4js/ui/N4JSUiModule.java
@@ -125,8 +125,8 @@ import org.eclipse.n4js.ui.workingsets.WorkingSetManagerBrokerImpl;
 import org.eclipse.n4js.ui.workingsets.WorkspaceRepositoriesProvider;
 import org.eclipse.n4js.utils.ProjectDescriptionLoader;
 import org.eclipse.n4js.utils.StatusHelper;
-import org.eclipse.n4js.utils.process.OutputStreamPrinterThreadProvider;
-import org.eclipse.n4js.utils.process.OutputStreamProvider;
+import org.eclipse.n4js.utils.io.OutputStreamPrinterThreadProvider;
+import org.eclipse.n4js.utils.io.OutputStreamProvider;
 import org.eclipse.n4js.utils.process.ProcessExecutor;
 import org.eclipse.n4js.utils.ui.editor.AvoidRefreshDocumentProvider;
 import org.eclipse.ui.plugin.AbstractUIPlugin;
