+++ plugins/org.eclipse.n4js.ui/src/org/eclipse/n4js/ui/internal/ContributingModule.java
@@ -68,8 +68,8 @@ import org.eclipse.n4js.utils.NodeModulesDiscoveryHelper;
 import org.eclipse.n4js.utils.ProjectDescriptionLoader;
 import org.eclipse.n4js.utils.StatusHelper;
 import org.eclipse.n4js.utils.WildcardPathFilterHelper;
-import org.eclipse.n4js.utils.process.OutputStreamPrinterThreadProvider;
-import org.eclipse.n4js.utils.process.OutputStreamProvider;
+import org.eclipse.n4js.utils.io.OutputStreamPrinterThreadProvider;
+import org.eclipse.n4js.utils.io.OutputStreamProvider;
 import org.eclipse.n4js.utils.process.ProcessExecutor;
 import org.eclipse.n4js.validation.helper.FolderContainmentHelper;
 import org.eclipse.xtext.builder.builderState.IBuilderState;
