+++ plugins/org.eclipse.n4js.utils/src/org/eclipse/n4js/utils/io/OutputStreamPrinterThread.java
@@ -8,7 +8,7 @@
  * Contributors:
  *   NumberFour AG - Initial API and implementation
  */
-package org.eclipse.n4js.utils.process;
+package org.eclipse.n4js.utils.io;
 
 import static com.google.common.base.Preconditions.checkNotNull;
